implementing search request
i’ve tried to make the search of a word that we give throug a form.
Then this form send the info to a php site that should make the
retrieving of the information from europeana. The problem is that i
don’t know exactly yet how to use that $_POST[“song”] into the command
line to get the europeana information. Since it is wroten in a script
and post is a php atribution i don’t know how to “paste” it in the line
that ask europeana the information. That is all happening in the <main>
tag in the search.php file.