Refresh button
I created a refresh button for the searchs because if we search 2 or 3
different things we will mix everything one after the otherone. i
though that you can just delete the meineListe2 delete but since jquery
has no speciffic order we can not say first show the information when
we click the button search and then with the second click of the search
button first delete the whole meineliste2 and then show the new
information that we searched. I’m still thinking about it