changes saved
i was in panic for few minutes but then i remembered that we are
working in github and you can’t make any mistakes without having the
oportunitty to go back :). For now the filters are not working yet but
i think i have an idea how they should work. language=en for english
for example and country=France for france. I still have to experiment a
bit with that but for tonight it was enough.